RESEARCH/.claude/services/ccpm-infra/src/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0
pydantic>=2.5.0
PyYAML>=6.0
